+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - rendering innate weapon attacks for PF2 characters
 - rendering damage reductions for PF2 characters
 
+### Modified
+- counting race weapon feats for PF2 characters
+
 ## [0.4.23] - 2026-04-05
 ### Added
 - creating additional spells for PF2 characters